### 2022-08-07 #4658 @editorialbot
[PRE REVIEW]: Accelerated oblique random survival forests
**Submitting author:** @bcjaeger (<a href="http://orcid.org/0000-0001-7399-2299">Byron Jaeger</a>)
**Repository:** https://github.com/bcjaeger/aorsf
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.1
