### 2022-08-24 #4658 @bcjaeger
Oh and in case I did not mention this, `aorsf` was reviewed at rOpenSci: https://github.com/ropensci/software-review/issues/532

I believe this qualifies `aorsf` for an expedited review by JOSS, depending on how the editor views the rOpenSci review.
