### 2022-08-24 #4658 @danielskatz
@bcjaeger and @Fei-Tao - because of the rOpenSci review, this indeed does not need to go through a full JOSS review. so I'll take it and handle the expedited review (and thanks to @Fei-Tao for agreeing to edit when we didn't know this)
