### 2022-08-24 #4658 @danielskatz
@bcjaeger - please add a statement of need section to the paper, and in that, or in the introduction, please explain more generally what this software does and what types of applications it has, remembering that JOSS has a very diverse readership, not all of which has detailed machine learning expertise. For example, the paper probably needs to explain what right-censored time-to-event data is and where it appears/is used, what survival decision trees are and where they appear/are used, what risk prediction models are and where they appear/are used, etc.  While you do this, I will do a bunch of the processing needed, including creating a review issue very shortly.
