### 2022-08-24 #4705 @danielskatz
And finally, for the first two references, please fix the cases of the journals to "Machine Learning" and "The Annals of Applied Statistics"
