### 2022-08-24 #4705 @bcjaeger
@danielskatz, thanks for the help today! Here is a list of updates based on your comments.

> Please add a statement of need section to the paper

Done, see lines 23-28 of the updated draft.

> in that, or in the introduction, please explain more generally what this software does and what types of applications it has, remembering that JOSS has a very diverse readership, not all of which has detailed machine learning expertise.

Good point. In the intro, I have added a general description of risk prediction (1st paragraph) and a summary of what `aorsf` does and what applications it has (second paragraph).

> the paper probably needs to explain what right-censored time-to-event data is and where it appears/is used, what survival decision trees are and where they appear/are used, what risk prediction models are and where they appear/are used, etc.

Agreed. In the update, I've included
- what risk prediction is and where it is used (lines 7-14)
- what censored outcomes are (lines 45-50; they are used in risk prediction, defined earlier)
- what survival decision trees are is included on lines 51-54

The bibliography and DOIs should also be in order now.
