### 2022-08-24 #4705 @bcjaeger
> Thanks - this looks good to me. I will also proofread this more carefully when you are ready - please let me know if that's now, and if not, when it is :)

Thank you! I would like to read through it once more tonight. May I add a post here sometime this evening (about 5 hours from now) to let you know when I've finished?
