### 2022-08-25 #4705 @bcjaeger
I think it is ready for your review. The only change I am considering is swapping the statement of need section with the background section so that the background section would come before the statement of need. This change may be helpful because the statement of need covers topics that are introduced in the background. I'm looking forward to hearing what you think.
