### 2022-08-25 #4705 @danielskatz
> I think it is ready for your review. The only change I am considering is swapping the statement of need section with the background section so that the background section would come before the statement of need. This change may be helpful because the statement of need covers topics that are introduced in the background. I'm looking forward to hearing what you think.

I'm not sure - I might just leave it the way it is, but it's up to you.

Edits:

* in line 32 and 37, "axis based" should be "axis-based"
* in line 38 and 40, "to axes" should be "to the axes"
* in line 46, "censored" perhaps should just be "censor"
* in line 82, I would change "negates them" to "it negates them"
* at the end of line 98, a period is needed

Let me know when you have made these changes, and decided about the text at the start, and I can proceed with the acceptance process.
