### 2022-08-25 #4705 @danielskatz
Note: aorsf was reviewed at rOpenSci: https://github.com/ropensci/software-review/issues/532

Thus, JOSS is not re-reviewing the software and is simply publishing it after ensuring that it also meets JOSS's requirements, including those for the paper itself.

See https://joss.readthedocs.io/en/latest/editing.html#processing-of-ropensci-reviewed-and-accepted-submissions for more on this expedited process.
