### 2022-09-22 #4705 @bcjaeger
> are any changes in the paper needed (based on the review process)?

Thanks for checking! I re-read the paper and can confirm it does not need an update.

> is the current software in zenodo or a similar repo now?

v0.0.2 is on CRAN now - is that what you mean by similar repo to zenodo? I am happy to publish it on zenodo if that would be helpful, just let me know.
