### 2022-09-22 683d73c
some changes from ropensci review
