+++ NEWS.md
@@ -1,5 +1,9 @@
 # aorsf 0.0.3
 
+* Added clarification to docs about pros/cons of different variable importance techniques
+
+* Added regression tests for aorsf versus obliqueRSF (they should be similar)
+
 * Additional support and tests for functions with long right hand sides
 
 * Updated out-of-bag vignette with more appropriate custom functions.
