+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   aorsf: aorsf.html
   oobag: oobag.html
   pd: pd.html
-last_built: 2022-09-21T15:40Z
+last_built: 2022-09-22T21:23Z
 urls:
   reference: https://bcjaeger.github.io/aorsf/reference
   article: https://bcjaeger.github.io/aorsf/articles
