+++ src/globals.h
@@ -71,6 +71,8 @@
  const PredType DEFAULT_PRED_TYPE = RISK;
  const int DEFAULT_N_SPLIT = 5;
 
+ const int VERBOSITY = 1;
+
  } // namespace aorsf
 
 #endif /* GLOBALS_H_ */
