### 2022-09-27 8c1ff32
tweaks to docs from ropensci rev
