### 2022-09-27 #4705 @bcjaeger
Done! here is the DOI: 10.5281/zenodo.7116855

I had to make a new release to get a zenodo DOI, and elected to use the tag of v0.0.3 since `aorsf` v0.0.3 incorporates all changes based on the rOpenSci review.
